WIP - trying to carry local address all the way through.
I'm abandoning this approach - it ends up quite messy, both
with the state-machine shenanigans and with trying to actually
get the address from the session into the connnection, when the
connection really only has the config in its hand.

Instead, let's give up on using a resolver, and just do a simple
inline numeric parse into the config and use that. Simples.